+++ src/api/mod.rs
@@ -48,19 +48,15 @@ use crate::platform::PeripheralId;
     derive(Serialize, Deserialize),
     serde(crate = "serde_cr")
 )]
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Default)]
 pub enum AddressType {
     Random,
+    #[default]
     Public,
 }
 
-impl Default for AddressType {
-    fn default() -> Self {
-        AddressType::Public
-    }
-}
-
 impl AddressType {
+    #[allow(clippy::should_implement_trait)]
     pub fn from_str(v: &str) -> Option<AddressType> {
         match v {
             "public" => Some(AddressType::Public),
@@ -96,6 +92,7 @@ pub struct ValueNotification {
 
 bitflags! {
     /// A set of properties that indicate what operations are supported by a Characteristic.
+    #[derive(Default, Debug, PartialEq, Eq, Ord, PartialOrd, Clone, Copy)]
     pub struct CharPropFlags: u8 {
         const BROADCAST = 0x01;
         const READ = 0x02;
@@ -105,12 +102,6 @@ bitflags! {
         const INDICATE = 0x20;
         const AUTHENTICATED_SIGNED_WRITES = 0x40;
         const EXTENDED_PROPERTIES = 0x80;
-    }
-}
-
-impl Default for CharPropFlags {
-    fn default() -> Self {
-        Self { bits: 0 }
     }
 }
 
@@ -144,6 +135,8 @@ pub struct Characteristic {
     /// supports. If you attempt an operation that is not supported by the characteristics (for
     /// example setting notify on one without the NOTIFY flag), that operation will fail.
     pub properties: CharPropFlags,
+    /// The descriptors of this characteristic.
+    pub descriptors: BTreeSet<Descriptor>,
 }
 
 impl Display for Characteristic {
@@ -156,8 +149,30 @@ impl Display for Characteristic {
     }
 }
 
+/// Add doc
+#[derive(Debug, Ord, PartialOrd, Eq, PartialEq, Clone)]
+pub struct Descriptor {
+    /// The UUID for this descriptor. This uniquely identifies its behavior.
+    pub uuid: Uuid,
+    /// The UUID of the service this descriptor belongs to.
+    pub service_uuid: Uuid,
+    /// The UUID of the characteristic this descriptor belongs to.
+    pub characteristic_uuid: Uuid,
+}
+
+impl Display for Descriptor {
+    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+        write!(f, "uuid: {:?}", self.uuid)
+    }
+}
+
 /// The properties of this peripheral, as determined by the advertising reports we've received for
 /// it.
+#[cfg_attr(
+    feature = "serde",
+    derive(Serialize, Deserialize),
+    serde(crate = "serde_cr")
+)]
 #[derive(Debug, Default, Clone)]
 pub struct PeripheralProperties {
     /// The address of this peripheral
@@ -178,8 +193,14 @@ pub struct PeripheralProperties {
     pub service_data: HashMap<Uuid, Vec<u8>>,
     /// Advertised services for this device
     pub services: Vec<Uuid>,
+    pub class: Option<u32>,
 }
 
+#[cfg_attr(
+    feature = "serde",
+    derive(Serialize, Deserialize),
+    serde(crate = "serde_cr")
+)]
 /// The filter used when scanning for BLE devices.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct ScanFilter {
@@ -264,6 +285,14 @@ pub trait Peripheral: Send + Sync + Clone + Debug {
     /// The stream will remain valid across connections and can be queried before any connection
     /// is made.
     async fn notifications(&self) -> Result<Pin<Box<dyn Stream<Item = ValueNotification> + Send>>>;
+
+    /// Write some data to the descriptor. Returns an error if the write couldn't be sent or (in
+    /// the case of a write-with-response) if the device returns an error.
+    async fn write_descriptor(&self, descriptor: &Descriptor, data: &[u8]) -> Result<()>;
+
+    /// Sends a read descriptor request to the device. Returns either an error if the request
+    /// was not accepted or the response from the device.
+    async fn read_descriptor(&self, descriptor: &Descriptor) -> Result<Vec<u8>>;
 }
 
 #[cfg_attr(