+++ src/droidplug/jni/objects.rs
@@ -10,7 +10,7 @@ use jni_utils::{future::JFuture, stream::JStream, uuid::JUuid};
 use std::{collections::HashMap, convert::TryFrom, iter::Iterator};
 use uuid::Uuid;
 
-use crate::api::{BDAddr, CharPropFlags, PeripheralProperties};
+use crate::api::{BDAddr, CharPropFlags, PeripheralProperties, ScanFilter};
 
 pub struct JPeripheral<'a: 'b, 'b> {
     internal: JObject<'a>,
@@ -22,6 +22,8 @@ pub struct JPeripheral<'a: 'b, 'b> {
     write: JMethodID<'a>,
     set_characteristic_notification: JMethodID<'a>,
     get_notifications: JMethodID<'a>,
+    read_descriptor: JMethodID<'a>,
+    write_descriptor: JMethodID<'a>,
     env: &'b JNIEnv<'a>,
 }
 
@@ -90,6 +92,16 @@ impl<'a: 'b, 'b> JPeripheral<'a, 'b> {
             "getNotifications",
             "()Lio/github/gedgygedgy/rust/stream/Stream;",
         )?;
+        let read_descriptor = env.get_method_id(
+            class,
+            "readDescriptor",
+            "(Ljava/util/UUID;Ljava/util/UUID;)Lio/github/gedgygedgy/rust/future/Future;",
+        )?;
+        let write_descriptor = env.get_method_id(
+            class,
+            "writeDescriptor",
+            "(Ljava/util/UUID;Ljava/util/UUID;[BI)Lio/github/gedgygedgy/rust/future/Future;",
+        )?;
         Ok(Self {
             internal: obj,
             connect,
@@ -100,6 +112,8 @@ impl<'a: 'b, 'b> JPeripheral<'a, 'b> {
             write,
             set_characteristic_notification,
             get_notifications,
+            read_descriptor,
+            write_descriptor,
             env,
         })
     }
@@ -233,6 +247,41 @@ impl<'a: 'b, 'b> JPeripheral<'a, 'b> {
             .l()?;
         JStream::from_env(self.env, stream_obj)
     }
+
+    pub fn read_descriptor(
+        &self,
+        characteristic: JUuid<'a, 'b>,
+        uuid: JUuid<'a, 'b>,
+    ) -> Result<JFuture<'a, 'b>> {
+        let future_obj = self
+            .env
+            .call_method_unchecked(
+                self.internal,
+                self.read_descriptor,
+                JavaType::Object("Lio/github/gedgygedgy/rust/future/Future;".to_string()),
+                &[characteristic.into(), uuid.into()],
+            )?
+            .l()?;
+        JFuture::from_env(self.env, future_obj)
+    }
+
+    pub fn write_descriptor(
+        &self,
+        characteristic: JUuid<'a, 'b>,
+        uuid: JUuid<'a, 'b>,
+        data: JObject<'a>,
+    ) -> Result<JFuture<'a, 'b>> {
+        let future_obj = self
+            .env
+            .call_method_unchecked(
+                self.internal,
+                self.write_descriptor,
+                JavaType::Object("Lio/github/gedgygedgy/rust/future/Future;".to_string()),
+                &[characteristic.into(), uuid.into(), data.into()],
+            )?
+            .l()?;
+        JFuture::from_env(self.env, future_obj)
+    }
 }
 
 pub struct JBluetoothGattService<'a: 'b, 'b> {
@@ -312,6 +361,7 @@ pub struct JBluetoothGattCharacteristic<'a: 'b, 'b> {
     get_uuid: JMethodID<'a>,
     get_properties: JMethodID<'a>,
     get_value: JMethodID<'a>,
+    get_descriptors: JMethodID<'a>,
     env: &'b JNIEnv<'a>,
 }
 
@@ -322,12 +372,14 @@ impl<'a: 'b, 'b> JBluetoothGattCharacteristic<'a, 'b> {
 
         let get_uuid = env.get_method_id(&class, "getUuid", "()Ljava/util/UUID;")?;
         let get_properties = env.get_method_id(&class, "getProperties", "()I")?;
+        let get_descriptors = env.get_method_id(&class, "getDescriptors", "()Ljava/util/List;")?;
         let get_value = env.get_method_id(&class, "getValue", "()[B")?;
         Ok(Self {
             internal: obj,
             get_uuid,
             get_properties,
             get_value,
+            get_descriptors,
             env,
         })
     }
@@ -371,6 +423,57 @@ impl<'a: 'b, 'b> JBluetoothGattCharacteristic<'a, 'b> {
             .l()?;
         jni_utils::arrays::byte_array_to_vec(self.env, value.into_inner())
     }
+
+    pub fn get_descriptors(&self) -> Result<Vec<JBluetoothGattDescriptor>> {
+        let obj = self
+            .env
+            .call_method_unchecked(
+                self.internal,
+                self.get_descriptors,
+                JavaType::Object("Ljava/util/List;".to_string()),
+                &[],
+            )?
+            .l()?;
+        let desc_list = JList::from_env(self.env, obj)?;
+        let mut desc_vec = vec![];
+        for desc in desc_list.iter()? {
+            desc_vec.push(JBluetoothGattDescriptor::from_env(self.env, desc)?);
+        }
+        Ok(desc_vec)
+    }
+}
+
+pub struct JBluetoothGattDescriptor<'a: 'b, 'b> {
+    internal: JObject<'a>,
+    get_uuid: JMethodID<'a>,
+    env: &'b JNIEnv<'a>,
+}
+
+impl<'a: 'b, 'b> JBluetoothGattDescriptor<'a, 'b> {
+    pub fn from_env(env: &'b JNIEnv<'a>, obj: JObject<'a>) -> Result<Self> {
+        let class = env.auto_local(env.find_class("android/bluetooth/BluetoothGattDescriptor")?);
+
+        let get_uuid = env.get_method_id(&class, "getUuid", "()Ljava/util/UUID;")?;
+        Ok(Self {
+            internal: obj,
+            get_uuid,
+            env,
+        })
+    }
+
+    pub fn get_uuid(&self) -> Result<Uuid> {
+        let obj = self
+            .env
+            .call_method_unchecked(
+                self.internal,
+                self.get_uuid,
+                JavaType::Object("Ljava/util/UUID;".to_string()),
+                &[],
+            )?
+            .l()?;
+        let uuid_obj = JUuid::from_env(self.env, obj)?;
+        Ok(uuid_obj.as_uuid()?)
+    }
 }
 
 pub struct JBluetoothDevice<'a: 'b, 'b> {
@@ -405,11 +508,49 @@ impl<'a: 'b, 'b> JBluetoothDevice<'a, 'b> {
     }
 }
 
+pub struct JScanFilter<'a> {
+    internal: JObject<'a>,
+}
+
+impl<'a> JScanFilter<'a> {
+    pub fn new(env: &'a JNIEnv<'a>, filter: ScanFilter) -> Result<Self> {
+        let uuids = env.new_object_array(
+            filter.services.len() as i32,
+            env.find_class("java/lang/String")?,
+            JObject::null(),
+        )?;
+        for (idx, uuid) in filter.services.into_iter().enumerate() {
+            let uuid_str = env.new_string(uuid.to_string())?;
+            env.set_object_array_element(uuids, idx as i32, uuid_str)?;
+        }
+        let obj = env.new_object(
+            JClass::from(
+                jni_utils::classcache::get_class(
+                    "com/nonpolynomial/btleplug/android/impl/ScanFilter",
+                )
+                .unwrap()
+                .as_obj(),
+            ),
+            //class.as_obj(),
+            "([Ljava/lang/String;)V",
+            &[uuids.into()],
+        )?;
+        Ok(Self { internal: obj })
+    }
+}
+
+impl<'a> From<JScanFilter<'a>> for JObject<'a> {
+    fn from(value: JScanFilter<'a>) -> Self {
+        value.internal
+    }
+}
+
 pub struct JScanResult<'a: 'b, 'b> {
     internal: JObject<'a>,
     get_device: JMethodID<'a>,
     get_scan_record: JMethodID<'a>,
     get_tx_power: JMethodID<'a>,
+    get_rssi: JMethodID<'a>,
     env: &'b JNIEnv<'a>,
 }
 
@@ -425,11 +566,13 @@ impl<'a: 'b, 'b> JScanResult<'a, 'b> {
             "()Landroid/bluetooth/le/ScanRecord;",
         )?;
         let get_tx_power = env.get_method_id(&class, "getTxPower", "()I")?;
+        let get_rssi = env.get_method_id(&class, "getRssi", "()I")?;
         Ok(Self {
             internal: obj,
             get_device,
             get_scan_record,
             get_tx_power,
+            get_rssi,
             env,
         })
     }
@@ -470,6 +613,17 @@ impl<'a: 'b, 'b> JScanResult<'a, 'b> {
             )?
             .i()
     }
+
+    pub fn get_rssi(&self) -> Result<jint> {
+        self.env
+            .call_method_unchecked(
+                self.internal,
+                self.get_rssi,
+                JavaType::Primitive(Primitive::Int),
+                &[],
+            )?
+            .i()
+    }
 }
 
 impl<'a: 'b, 'b> TryFrom<JScanResult<'a, 'b>> for (BDAddr, Option<PeripheralProperties>) {
@@ -504,11 +658,16 @@ impl<'a: 'b, 'b> TryFrom<JScanResult<'a, 'b>> for (BDAddr, Option<PeripheralProp
                 None
             } else {
                 let device_name_str = JavaStr::from_env(result.env, device_name_obj)?;
+                // On Android, there is a chance that a device name may not actually be valid UTF-8.
+                // We're given the full buffer, regardless of if it's just UTF-8 characters,
+                // possibly c str with null characters, or whatever. We should try UTF-8 first, if
+                // that doesn't work out, see if there's a null termination character in it and try
+                // parsing that.
                 Some(
-                    device_name_str
-                        .to_str()
-                        .map_err(|e| Self::Error::Other(e.into()))?
-                        .to_string(),
+                    String::from_utf8_lossy(device_name_str.to_bytes())
+                        .chars()
+                        .filter(|&c| c != '\u{fffd}')
+                        .collect(),
                 )
             };
 
@@ -519,6 +678,8 @@ impl<'a: 'b, 'b> TryFrom<JScanResult<'a, 'b>> for (BDAddr, Option<PeripheralProp
             } else {
                 Some(tx_power_level as i16)
             };
+
+            let rssi = Some(result.get_rssi()? as i16);
 
             let manufacturer_specific_data_array = record.get_manufacturer_specific_data()?;
             let manufacturer_specific_data_obj: &JObject = &manufacturer_specific_data_array;
@@ -576,7 +737,8 @@ impl<'a: 'b, 'b> TryFrom<JScanResult<'a, 'b>> for (BDAddr, Option<PeripheralProp
                 manufacturer_data,
                 service_data,
                 services,
-                rssi: None,
+                rssi,
+                class: None,
             })
         };
         Ok((addr, properties))