+++ src/winrtble/ble/device.rs
@@ -17,7 +17,7 @@ use windows::{
     Devices::Bluetooth::{
         BluetoothCacheMode, BluetoothConnectionStatus, BluetoothLEDevice,
         GenericAttributeProfile::{
-            GattCharacteristic, GattCommunicationStatus, GattDeviceService,
+            GattCharacteristic, GattCommunicationStatus, GattDescriptor, GattDeviceService,
             GattDeviceServicesResult,
         },
     },
@@ -29,6 +29,7 @@ pub type ConnectedEventHandler = Box<dyn Fn(bool) + Send>;
 pub struct BLEDevice {
     device: BluetoothLEDevice,
     connection_token: EventRegistrationToken,
+    services: Vec<GattDeviceService>,
 }
 
 impl BLEDevice {
@@ -59,6 +60,7 @@ impl BLEDevice {
         Ok(BLEDevice {
             device,
             connection_token,
+            services: vec![],
         })
     }
 
@@ -98,19 +100,53 @@ impl BLEDevice {
         let async_result = service
             .GetCharacteristicsWithCacheModeAsync(BluetoothCacheMode::Uncached)?
             .await?;
+
+        match async_result.Status() {
+            Ok(GattCommunicationStatus::Success) => {
+                let results = async_result.Characteristics()?;
+                debug!("characteristics {:?}", results.Size());
+                Ok(results.into_iter().collect())
+            }
+            Ok(GattCommunicationStatus::ProtocolError) => Err(Error::Other(
+                format!(
+                    "get_characteristics for {:?} encountered a protocol error",
+                    service
+                )
+                .into(),
+            )),
+            Ok(status) => {
+                debug!("characteristic read failed due to {:?}", status);
+                Ok(vec![])
+            }
+            Err(e) => Err(Error::Other(
+                format!("get_characteristics for {:?} failed: {:?}", service, e).into(),
+            )),
+        }
+    }
+
+    pub async fn get_characteristic_descriptors(
+        characteristic: &GattCharacteristic,
+    ) -> Result<Vec<GattDescriptor>> {
+        let async_result = characteristic
+            .GetDescriptorsWithCacheModeAsync(BluetoothCacheMode::Uncached)?
+            .await?;
         let status = async_result.Status();
         if status == Ok(GattCommunicationStatus::Success) {
-            let results = async_result.Characteristics()?;
-            debug!("characteristics {:?}", results.Size());
+            let results = async_result.Descriptors()?;
+            debug!("descriptors {:?}", results.Size());
             Ok(results.into_iter().collect())
         } else {
             Err(Error::Other(
-                format!("get_characteristics for {:?} failed: {:?}", service, status).into(),
+                format!(
+                    "get_characteristic_descriptors for {:?} failed: {:?}",
+                    characteristic, status
+                )
+                .into(),
             ))
         }
     }
 
-    pub async fn discover_services(&self) -> Result<Vec<GattDeviceService>> {
+    pub async fn discover_services(&mut self) -> Result<&[GattDeviceService]> {
         let winrt_error = |e| Error::Other(format!("{:?}", e).into());
         let service_result = self.get_gatt_services(BluetoothCacheMode::Cached).await?;
         let status = service_result.Status().map_err(winrt_error)?;
@@ -122,10 +158,10 @@ impl BLEDevice {
                 .map_err(winrt_error)?
                 .into_iter()
                 .collect();
-            debug!("services {:?}", services.len());
-            return Ok(services);
+            self.services = services;
+            debug!("services {:?}", self.services.len());
         }
-        Ok(Vec::new())
+        Ok(self.services.as_slice())
     }
 }
 
@@ -137,6 +173,12 @@ impl Drop for BLEDevice {
         if let Err(err) = result {
             debug!("Drop:remove_connection_status_changed {:?}", err);
         }
+
+        self.services.iter().for_each(|service| {
+            if let Err(err) = service.Close() {
+                debug!("Drop:remove_gatt_Service {:?}", err);
+            }
+        });
 
         let result = self.device.Close();
         if let Err(err) = result {